Refactoring code, Settings and Environment Variables
