+++ app/dependencies/database.py
@@ -1,4 +1,4 @@
-from db.dev import SessionLocal, engine, Base
+from db.connection import SessionLocal, engine, Base
 
 from models.comuna import ComunaModel
 from models.barrio import BarrioModel
