+++ app/dependencies/settings.py
@@ -1,0 +1,9 @@
+from functools import lru_cache
+
+from config import Settings
+
+
+# Dependency
+@lru_cache()
+def get_settings():
+    return Settings()
