+++ app/routers/areaspro.py
@@ -1,7 +1,7 @@
 from fastapi import Depends, APIRouter, HTTPException, status
 from sqlalchemy.orm import Session
 
-from dependencies.db import get_db
+from dependencies.database import get_db
 from schemas.areapro import AreaProSchema, AreaProCreate, AreaProUpdate
 from crud.areapro import AreaProCRUD
 from schemas.tangara import TangaraSchema
